rework package and namespace references
up to now namespace do not have an reference to other namespaces. Now each namespace store a reference to all namespaces that are visible in the correpsonding fidl file.

package store only references to other packages but not to istelf

The namespaces of all used types within a fidl file must be imported. Transitive dependency is not supported.