add unit test to show an package import bug
Currently a package has refernces to all imported namespaces. The processor does not cosnider the import directive. A import directive defines which namespace is visible to all namespaces define in the current file not in the current package.

In this unittest the package references point to different objects altough there is just one package.

Fix:
Each namespace has a list of namespace references of all visible namespaces.